Don't log secret of s3. (#292)
* backup: don't log secret of s3.

* br_s3: always print log.

* *: remove query string on logging arguments.

* br_s3: reset BR_LOG_TO_TERM on fail.

* Apply suggestions from code review


* Apply suggestions from code review


* *: remove query string on logging arguments.

+++ tests/br_s3/run.sh
@@ -58,7 +58,19 @@ done
 
 # backup full
 echo "backup start..."
-run_br --pd $PD_ADDR backup full -s "s3://mybucket/$DB?endpoint=http://$S3_ENDPOINT"
+BACKUP_LOG="backup.log"
+rm -f $BACKUP_LOG
+unset BR_LOG_TO_TERM
+run_br --pd $PD_ADDR backup full -s "s3://mybucket/$DB?endpoint=http://$S3_ENDPOINT" \
+    --log-file $BACKUP_LOG || \
+    ( cat $BACKUP_LOG && BR_LOG_TO_TERM=1 && exit 1 )
+cat $BACKUP_LOG
+BR_LOG_TO_TERM=1
+
+if grep -i $MINIO_SECRET_KEY $BACKUP_LOG; then
+    echo "Secret key logged in log. Please remove them."
+    exit 1
+fi
 
 for i in $(seq $DB_COUNT); do
     run_sql "DROP DATABASE $DB${i};"
@@ -66,7 +78,19 @@ done
 
 # restore full
 echo "restore start..."
-run_br restore full -s "s3://mybucket/$DB" --pd $PD_ADDR --s3.endpoint="http://$S3_ENDPOINT"
+RESTORE_LOG="restore.log"
+rm -f $RESTORE_LOG
+unset BR_LOG_TO_TERM
+run_br restore full -s "s3://mybucket/$DB" --pd $PD_ADDR --s3.endpoint="http://$S3_ENDPOINT" \
+    --log-file $RESTORE_LOG || \
+    ( cat $RESTORE_LOG && BR_LOG_TO_TERM=1 && exit 1 )
+cat $RESTORE_LOG
+BR_LOG_TO_TERM=1
+
+if grep -i $MINIO_SECRET_KEY $RESTORE_LOG; then
+    echo "Secret key logged in log. Please remove them."
+    exit 1
+fi
 
 for i in $(seq $DB_COUNT); do
     row_count_new[${i}]=$(run_sql "SELECT COUNT(*) FROM $DB${i}.$TABLE;" | awk '/COUNT/{print $2}')
